Refactor component imports to use absolute paths, remove unused components, and update text for consistency. Added new PDFViewer and ProgrammerVector components.
